refactor: retire core technical debt
