+++ .github/workflows/ci.yml
@@ -57,6 +57,11 @@ jobs:
       - name: Generate checked build inputs
         run: task generate
 
+      - name: Check committed generated artifacts
+        run: |
+          changed="$(git status --porcelain -- .env.example docs/configuration.md internal/config/config_gen.go internal/configspec/names_gen.go schemas web/generated)"
+          test -z "$changed" || { printf 'committed generated artifacts are stale:\n%s\n' "$changed" >&2; exit 1; }
+
       - name: Build browser assets
         run: task build
 
@@ -119,9 +124,13 @@ jobs:
               bun run test:theme
               bun run test:datastar-lit
               bun run test:app-shell
+              bun run typecheck:app
+              bun run typecheck:contracts
+              bun run test:production-assets
+              bun run test:movielens-performance-harness
           - name: reports
             tests: |
-              bun run test:table-selection
+              bun run test:interaction-selection
               bun run test:record-table
               bun run test:visual-modal
               bun run test:catalog-page
@@ -144,6 +153,8 @@ jobs:
               bun run test:login-page
               bun run test:asset-lineage
               bun run test:semantic-model-graph
+              bun run test:datastar-inspector
+              bun run test:topology-background
 
     steps:
       - name: Checkout
@@ -168,6 +179,34 @@ jobs:
 
       - name: Run frontend tests
         run: ${{ matrix.tests }}
+
+  go-analysis:
+    name: Go static and race analysis
+    runs-on: ubuntu-24.04
+    timeout-minutes: 25
+    needs: prepare
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7
+
+      - name: Set up Go
+        uses: actions/setup-go@924ae3a1cded613372ab5595356fb5720e22ba16 # v6
+        with:
+          go-version-file: go.mod
+          cache: true
+
+      - name: Download generated assets
+        uses: actions/download-artifact@3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c # v8.0.1
+        with:
+          name: generated-assets
+          path: .
+
+      - name: Run Go static analysis
+        run: go vet ./...
+
+      - name: Run race-sensitive package tests
+        run: go test -race ./pkg/... ./internal/access ./internal/runtimehost
 
   ui-route-qa:
     name: UI route QA
