+++ internal/access/http/audit_test.go
@@ -1,0 +1,38 @@
+package http
+
+import (
+	"context"
+	stdhttp "net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/Yacobolo/libredash/internal/access"
+)
+
+type auditedMutationRepository struct {
+	access.Repository
+	called bool
+}
+
+func (r *auditedMutationRepository) RunAuditedMutation(ctx context.Context, mutation func(access.Repository) (access.AuditEventInput, error)) error {
+	r.called = true
+	_, err := mutation(r.Repository)
+	return err
+}
+
+func TestRunAuditedMutationUsesRepositoryTransaction(t *testing.T) {
+	repo := &auditedMutationRepository{}
+	request := httptest.NewRequest(stdhttp.MethodPost, "/", nil)
+	mutationCalled := false
+
+	err := runAuditedMutation(request, repo, func(access.Repository) (access.AuditEventInput, error) {
+		mutationCalled = true
+		return access.AuditEventInput{Action: "grant.created"}, nil
+	})
+	if err != nil {
+		t.Fatalf("run audited mutation: %v", err)
+	}
+	if !repo.called || !mutationCalled {
+		t.Fatalf("transaction called = %v, mutation called = %v", repo.called, mutationCalled)
+	}
+}
