+++ web/components/shared/toon-language.ts
@@ -108,4 +108,4 @@ export const toonLanguage = {
       ],
     },
   },
-} as const
+}
